+++ db/db_role.py
@@ -1,11 +1,14 @@
 
 
 from schemas import RoleBase
-from db.models import DbRole
+from db.models import DbRole,DbEmployee
 from sqlalchemy.orm import Session
 from fastapi import HTTPException,status
+from utils.string_utils import normalize_string
 
 def create_role(db: Session, request: RoleBase):
+    
+
     if not request.role_code or not request.role_code.strip() or request.role_code.strip().lower() == "string":
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
@@ -18,20 +21,26 @@ def create_role(db: Session, request: RoleBase):
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Role name cannot be empty or default value"
     )
-    if db.query(DbRole).filter(
-        DbRole.role_code.ilike(request.role_code.strip())).first():
-        raise HTTPException(
+    normalized_code = normalize_string(request.role_code)
+    normalized_name = normalize_string(request.role_name)
+
+
+    if db.query(DbRole).filter(DbRole.role_code.ilike(normalized_code)).first():
+       raise HTTPException(
         status_code=status.HTTP_400_BAD_REQUEST,
         detail="Role code already exists"
     )
 
 
-    if db.query(DbRole).filter(
-       DbRole.role_name.ilike(request.role_name.strip())).first():
-       raise HTTPException(
-        status_code=status.HTTP_400_BAD_REQUEST,
-        detail="Role name already exists"
-    )
+    existing_roles = db.query(DbRole).all()
+    for role in existing_roles:
+       if normalize_string(role.role_name) == normalized_name:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Role name already exists"
+        )
+
+    
 
     new_role = DbRole(
         role_code=request.role_code,
@@ -72,7 +81,25 @@ def update_role(db: Session, id: int, request: RoleBase):
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Role name cannot be empty or default value"
-        )    
+        ) 
+    normalized_code = normalize_string(request.role_code)
+    normalized_name = normalize_string(request.role_name)
+
+
+    if db.query(DbRole).filter(DbRole.role_code.ilike(normalized_code)).first():
+       raise HTTPException(
+        status_code=status.HTTP_400_BAD_REQUEST,
+        detail="Role code already exists"
+    )
+
+
+    existing_roles = db.query(DbRole).all()
+    for role in existing_roles:
+       if normalize_string(role.role_name) == normalized_name:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Role name already exists"
+        )       
     role.role_code = request.role_code
     role.role_name = request.role_name
 
@@ -92,6 +119,14 @@ def delete_role(db:Session,id:int):
             status_code=status.HTTP_404_NOT_FOUND, 
             detail=f"Role with id={id} not found"
         )
+    employee_count = db.query(DbEmployee).filter(
+        DbEmployee.role_id == id).count()
+
+    if employee_count > 0:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This role has employees. Move or delete them first."
+        )    
     db.delete(role)
     db.commit()
     return 'ok'
